Add arguments of type CompleteHeadVisitor<uint8> and PartialHeadVisitor<uint8> to function "visit" of the class IHead.

+++ cpp/subprojects/common/src/mlrl/common/rule_refinement/prediction_complete.cpp
@@ -7,7 +7,7 @@
 
 static inline std::unique_ptr<IHead> createHeadInternally(const CompletePrediction<uint8>& prediction) {
     uint32 numElements = prediction.getNumElements();
-    std::unique_ptr<CompleteHead<float32>> headPtr = std::make_unique<CompleteHead<float32>>(numElements);
+    std::unique_ptr<CompleteHead<uint8>> headPtr = std::make_unique<CompleteHead<uint8>>(numElements);
     util::copyView(prediction.values_cbegin(), headPtr->values_begin(), numElements);
     return headPtr;
 }
